+++ app/src/core/execution/index.ts
@@ -1,2 +1,3 @@
 export * from './executionTypes'
 export * from './executionStore'
+export * from './executionLifecycle'